# api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This api powers automart services.  You can find out more about     automart [here](https://stepheng323.github.io/Automart/UI). For protected route, get a token by either the sign up route or sign in route, click on the authorized botton and add your token followin this pattern `bearer: token goes here`",
		"version": "1.0.0",
		"title": "Automart",
		"termsOfService": "",
		"contact": {
			"email": "stepheng323@gmail.com"
		},
		"license": {
			"name": "ISC",
			"url": "https://www.isc.org"
		}
	},
	"host": "automartendpoints.herokuapp.com",
	"basePath": "/api/v1",
...
	"schemes": ["https"],
	"paths": {
...
		"/car?status=available": {
			"get": {
				"tags": ["car"],
				"summary": "Get all unsold cars",
				"description": "Get all cars whose status is available",
				"operationId": "available",
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/Car"
						}
					},
					"400": {
						"description": "No car found"
					}
				},
				"security": [
					{
						"Bearer": []
					}
				]
			}
		},
...
	},
	"securityDefinitions": {
		"Bearer": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"definitions": {
...
		"Car": {
			"type": "object",
			"properties": {
				"state": {
					"type": "string"
				},
				"status": {
					"type": "string"
				},
				"price": {
					"type": "number"
				},
				"manufacturer": {
					"type": "string"
				},
				"model": {
					"type": "string"
				},
				"body_type": {
					"type": "string"
				},
				"product_image": {
					"type": "file"
				}
			}
		}
	}
}